pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "video-to-powerpoint"
version = "2.0.0"
description = "Capture video presentations and convert them to PowerPoint slides automatically"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Derek Plemons", email = "plemons.derek@gmail.com"}
]
keywords = ["powerpoint", "screenshot", "presentation", "video", "capture", "slides"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Environment :: Win32 (MS Windows)",
    "Intended Audience :: End Users/Desktop",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics :: Capture :: Screen Capture",
    "Topic :: Office/Business :: Office Suites",
]
requires-python = ">=3.9"
dependencies = [
    "mss>=9.0.0",
    "pynput>=1.7.6",
    "Pillow>=10.0.0",
    "python-pptx>=0.6.21",
    "scikit-image>=0.22.0",
    "customtkinter>=5.2.0",
    "img2pdf>=0.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "pyinstaller>=6.0.0",
]

[project.urls]
Homepage = "https://github.com/dplem/Video-to-PowerPoint"
Repository = "https://github.com/dplem/Video-to-PowerPoint"
Issues = "https://github.com/dplem/Video-to-PowerPoint/issues"

[project.scripts]
video-to-ppt = "video_to_powerpoint.cli:main"

[project.gui-scripts]
video-to-ppt-gui = "video_to_powerpoint.gui:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["video_to_powerpoint*"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by black)
]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=video_to_powerpoint --cov-report=term-missing"
